% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapperMethods.R
\name{calculateConsensus}
\alias{calculateConsensus}
\title{Wrapper function to calculate consensus}
\usage{
calculateConsensus(
  list_markers,
  input_matrix_train,
  clusters_train,
  num_markers = 15,
  method = "other",
  list_weight_num = list_weight_num,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{list_markers}{List of string vector, with each element of list corresponding to a marker method}

\item{input_matrix_train}{Training feature matrix (cells as rows, features as columns)}

\item{clusters_train}{Training cluster annotation vector}

\item{num_markers}{Number of markers}

\item{method}{Consensus method, based on the methods in `list_markers`
\itemize{
  \item \code{other} (default) Consensus calculated based on majority rule.
  \item \code{weighted}: The contribution of each marker is weighted according `list_weight_num`
  \item \code{fstat}: The contribution of each marker is weighted according to the fstat algorithm.
  \item \code{xgBoost}: The contribution of each marker is weighted according xgBoost.
}}

\item{list_weight_num}{Weight of each marker. Only used if `method` is weighted.}
}
\value{
Consensus markers
}
\description{
Wrapper function to calculate consensus
}
