% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapperPerformance.R
\name{calculatePerformance}
\alias{calculatePerformance}
\title{Wrapper function for geneBasisPerformance}
\usage{
calculatePerformance(test_prediction, unique_clusters_sample, ...)
}
\arguments{
\item{test_prediction}{Selected markers}

\item{unique_clusters_sample}{Unique clusters}
}
\value{
List of two elements: performance of each cluster, and performance over all clusters. Metrics include precision_weighted, precision_macro, recall_weighted, recall_macro, F1_macro, F1_weighted, precision_micro
}
\description{
Wrapper function for geneBasisPerformance
}
