% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findClusterMarkers.R
\name{findClusterMarkers}
\alias{findClusterMarkers}
\title{Find the most informative cluster markers}
\usage{
findClusterMarkers(input_matrix, clusters, num_markers = 15, method = "all")
}
\arguments{
\item{input_matrix}{Feature matrix with cells as columns, and features as rows.}

\item{clusters}{Cluster annotation for each cell.}

\item{num_markers}{Number of markers to output.}

\item{method}{List of methods to find cluster markers.
\itemize{
  \item \code{citeFUSE}
  \item \code{sc2marker}
  \item \code{geneBasis}
  \item \code{all}: Use all methods
}}
}
\value{
A list containing
\itemize{
  \item \code{markers}: Informative markers for each method
  \item \code{score_min}: Minimum of cell type scores as predicted by geneBasis
  \item \code{score_median}: Median of cell type scores as predicted by geneBasis
  \item \code{celltype_stat}: Cell type stats as predicted by geneBasis
}
}
\description{
Find the most informative cluster markers
}
