% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findClusterPerformance.R
\name{findClusterPerformance}
\alias{findClusterPerformance}
\title{Find the most informative cluster markers}
\usage{
findClusterPerformance(
  matrix_all,
  clusters,
  clusters_sel = "ALLCLUSTER",
  num_markers,
  subsample_num,
  train_test_ratio = 0.9,
  method_cluster = "all",
  method_performance = "all",
  cluster_proportion = "proportional",
  verbose = FALSE,
  ...
)
}
\arguments{
\item{matrix_all}{Feature matrix with cells as rows, and features as columns.}

\item{clusters}{Cluster annotation for each cell.}

\item{clusters_sel}{Selected clusters}

\item{num_markers}{Number of markers to output.}

\item{subsample_num}{Number of cells to sub-sample}

\item{train_test_ratio}{Ratio of training to test data}

\item{method_cluster}{List of methods to find cluster markers.
\itemize{
  \item \code{citeFUSE}
  \item \code{sc2marker}
  \item \code{geneBasis}
  \item \code{xgBoost}
  \item \code{all}: Use all methods
}}

\item{method_performance}{List of methods to evaluate performance.
\itemize{
  \item \code{xgBoost}
  \item \code{geneBasis}
  \item \code{all}: Use all methods
}}

\item{cluster_proportional}{Sub-sampling equally or proportional
\itemize{
  \item \code{equal}
  \item \code{proportional}
}}
}
\value{
A list containing
\itemize{
  \item \code{markers}: Informative markers for each method
  \item \code{performance}: Performance of each method
}
}
\description{
Find the most informative cluster markers
}
