% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapperMethods.R
\name{fstatWrapper}
\alias{fstatWrapper}
\title{Wrapper function for fstat}
\usage{
fstatWrapper(input_matrix, clusters, num_markers, ...)
}
\arguments{
\item{input_matrix}{Marker matrix with cells as rows, and features as columns.}

\item{clusters}{Cell type annotation}

\item{num_markers}{Number of markers to output}
}
\value{
The most informative markers determined by fstat
}
\description{
Wrapper function for fstat
}
