% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapperPerformance.R
\name{geneBasisPredictCell}
\alias{geneBasisPredictCell}
\title{Wrapper function for geneBasisPerformance}
\usage{
geneBasisPredictCell(
  markers_sel,
  input_matrix_test,
  clusters_test,
  unique_clusters_sample,
  ...
)
}
\arguments{
\item{markers_sel}{Selected markers}

\item{input_matrix_test}{Feature test matrix with cells as columns, and features as rows.}

\item{clusters_test}{Cluster annotation for test set}

\item{unique_clusters_sample}{Unique clusters}
}
\value{
True positive rate calculated using get_celltype_mapping from geneBasisR
}
\description{
Wrapper function for geneBasisPerformance
}
