% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapperMethods.R
\name{geneBasisWrapper}
\alias{geneBasisWrapper}
\title{Wrapper function for geneBasis}
\usage{
geneBasisWrapper(sce, clusters, num_markers = 15, ...)
}
\arguments{
\item{sce}{Single cell experiment object}

\item{num_markers}{Number of markers to output}
}
\value{
The most informative markers determined by geneBasis
}
\description{
Wrapper function for geneBasis
}
