% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performanceAllMarkers.R
\name{performanceAllMarkers}
\alias{performanceAllMarkers}
\title{Find performance measure of predicted markers}
\usage{
performanceAllMarkers(
  list_markers,
  final_out,
  method = "xgBoost",
  nrounds = 1500,
  nthread = 6,
  testSet = "test",
  verbose = FALSE,
  ...
)
}
\arguments{
\item{list_markers}{Output of `findClusterMarkers` function.}

\item{final_out}{Output of `processSubsampling` function.}

\item{method}{List of methods to find the performance of the cluster markers.
\itemize{
  \item \code{xgBoost}
  \item \code{geneBasis}
  \item \code{all}: Use all methods
}}

\item{nrounds}{Number of rounds used in xgBoost algorithm.}

\item{nthread}{Number of threads used in xgBoost algorithm.}
}
\value{
A list containing the performance of testing set, for each marker identification method.
}
\description{
Find performance measure of predicted markers
}
