% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapperPerformance.R
\name{performanceMarkers}
\alias{performanceMarkers}
\title{Find performance measure of predicted markers}
\usage{
performanceMarkers(
  markers_sel,
  input_matrix_train,
  input_matrix_test,
  unique_clusters_sample,
  clusters_num_train,
  clusters_num_test,
  clusters_train,
  clusters_test,
  method = "all",
  nrounds = 1500,
  nthread = 6,
  method_marker_name = "xgBoost",
  verbose = FALSE,
  ...
)
}
\arguments{
\item{markers_sel}{Selected markers for classification}

\item{input_matrix_train}{Feature training matrix with cells as columns, and features as rows.}

\item{input_matrix_test}{Feature test matrix with cells as columns, and features as rows.}

\item{unique_clusters_sample}{Unique clusters}

\item{clusters_num_train}{Cluster annotation for training set (numerical)}

\item{clusters_num_test}{Cluster annotation for testing set (numerical)}

\item{clusters_train}{Cluster annotation for training set}

\item{clusters_test}{Cluster annotation for test set}

\item{method}{List of methods to find cluster markers.
\itemize{
  \item \code{xgBoost}
  \item \code{geneBasis}
  \item \code{all}: Use all methods
}}

\item{nrounds}{Number of rounds used in xgBoost algorithm.}

\item{nthread}{Number of threads used in xgBoost algorithm.}
}
\value{
A list containing performance of testing set, for each cluster
}
\description{
Find performance measure of predicted markers
}
