% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performanceOwnMarkers.R
\name{performanceOwnMarkers}
\alias{performanceOwnMarkers}
\title{Find performance measure of the user's own markers}
\usage{
performanceOwnMarkers(
  user_markers,
  final_out,
  method = "all",
  nrounds = 1500,
  nthread = 6,
  testSet = "test",
  verbose = FALSE,
  ...
)
}
\arguments{
\item{user_markers}{Markers input from the users.}

\item{final_out}{Output of `processSubsampling` function.}

\item{method}{List of methods to find the performance of the cluster markers.
\itemize{
  \item \code{xgBoost}
  \item \code{geneBasis}
  \item \code{all}: Use all methods
}}

\item{nrounds}{Number of rounds used in xgBoost algorithm.}

\item{nthread}{Number of threads used in xgBoost algorithm.}
}
\value{
A list containing the performance of testing set, for user's own marker panel.
}
\description{
Find performance measure of the user's own markers
}
