% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotExpression.R
\name{plotExpression}
\alias{plotExpression}
\title{Visualize gene expression data for a list of markers using UMAP or violin plot}
\usage{
plotExpression(
  list_markers,
  input_obj,
  plot_type = "umap",
  markers_plot = NA,
  order_method = NA,
  cluster_levels = NA,
  marker_name_plot = TRUE,
  return_ggplot = FALSE,
  title_size = 10,
  text_size = 10,
  point_size = 1,
  ...
)
}
\arguments{
\item{list_markers}{A list of gene markers to be plotted (required).}

\item{input_obj}{An input object containing gene expression data and cluster information (required).}

\item{plot_type}{The type of plot to generate, either "umap" or "violin" (default is "umap").}

\item{markers_plot}{A vector of markers to be plotted (default is NA, which will use all markers in the `list_markers` object).}

\item{order_method}{A method for ordering the markers (default is alphabetical).}

\item{cluster_levels}{A vector of cluster levels (default will use the unique clusters in the input object).}

\item{marker_name_plot}{A logical value indicating whether to display marker names on the plot (default is TRUE).}

\item{return_ggplot}{A logical value indicating whether to return a ggplot object (default is FALSE).}

\item{title_size}{The size of the title text (default is 30).}

\item{text_size}{The size of the text (default is 30).}

\item{point_size}{The size of the points in the plot (default is 3).}

\item{...}{Additional arguments passed to the underlying plotting functions.}
}
\value{
If return_ggplot is TRUE, a ggplot object with the generated plot.
}
\description{
Visualize gene expression data for a list of markers using UMAP or violin plot
}
\examples{
# Given an input object with gene expression data and cluster information, and a list of markers:
plotExpression(list_markers, input_obj, plot_type = "umap")

}
