% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapperPlots.R
\name{plotMarkers}
\alias{plotMarkers}
\title{Plot markers identified by each method on a heatmap}
\usage{
plotMarkers(
  list_markers,
  text_size = 20,
  color_tile = "blue",
  fill_tile = "blue",
  coord_ratio = 0.2,
  ...
)
}
\arguments{
\item{list_markers}{List of selected markers from each method, from output of findClusterMarkers function.}

\item{text_size}{Text size}

\item{color_tile}{Color of each tile}

\item{fill_tile}{Fill of each tile}

\item{coord_ratio}{Heatmap tile coordinate ratio}
}
\value{
A ggplot output
}
\description{
Plot markers identified by each method on a heatmap
}
