% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapperPlots.R
\name{plotPerformance}
\alias{plotPerformance}
\title{Plot performance of each marker}
\usage{
plotPerformance(
  list_performance,
  metric = "F1",
  tile_text_size = 3,
  text_size = 8,
  ...
)
}
\arguments{
\item{list_performance}{List of performance metrics, from output of performanceAllMarkers function.}

\item{tile_text_size}{Tile text size}

\item{text_size}{Text size}
}
\value{
A list of ggplot outputs
}
\description{
Plot performance of each marker
}
