% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapperProcessData.R
\name{processClusterSelection}
\alias{processClusterSelection}
\title{Select clusters to compare with (based on user input)}
\usage{
processClusterSelection(
  sc_out,
  clusters_sel = "all_clusters",
  verbose = TRUE,
  ...
)
}
\arguments{
\item{sc_out}{Output of function `processInputFormat`.}

\item{clusters_sel}{Subset of clusters to identify. Default is to use all clusters.}
}
\value{
List containing
\itemize{
  \item \code{matrix}: Feature matrix (cells as rows, features as columns)
  \item \code{clusters}: Cluster annotation vector
}
}
\description{
Select clusters to compare with (based on user input)
}
