% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapperProcessData.R
\name{processInputFormat}
\alias{processInputFormat}
\title{Process input format (feature matrix, Seurat object, SCE object) to a format usable in other functions}
\usage{
processInputFormat(
  sc_object,
  clusters_all = NULL,
  sce_cluster = NULL,
  seurat_assay = NULL,
  seurat_slot = NULL,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{sc_object}{Input matrix and cell type object (required). Can be
\itemize{
  \item \code{Matrix:} Feature matrix (cells as rows, features as columns). cluster annotation has to be provided in `clusters_all` input.
  \item \code{SCE object}: SCE object, with the cluster annotation specified by the `sce_cluster` column name in the `colData` slot.
  \item \code{Seurat object}: Seurat object, with the cluster annotation in the `active.idents` slot.
}}

\item{clusters_all}{Cluster annotation, used if the input `sc_object` is a feature matrix.}

\item{sce_cluster}{Cluster annotation column name in the 'colData' slot, to be used if the input `sc_object` is a SCE object.}

\item{seurat_assay}{Seurat assay slot, if the input is a Seurat object. If not specified, default assay is used.}

\item{seurat_slot}{Seurat data slot, if the input is a Seurat object. If not specified, the `data` slot will be used.}
}
\value{
List containing
\itemize{
  \item \code{matrix}: Feature matrix (cells as rows, features as columns)
  \item \code{clusters}: Cluster annotation vector
}
}
\description{
Process input format (feature matrix, Seurat object, SCE object) to a format usable in other functions
}
