% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapperProcessData.R
\name{processSubsampling}
\alias{processSubsampling}
\title{Sub-sample and split data into training, validation and test set}
\usage{
processSubsampling(
  cluster_selection_out,
  subsample_num = 1000,
  cluster_proportion = "proportional",
  verbose = TRUE,
  seed = 8,
  ...
)
}
\arguments{
\item{cluster_selection_out}{Output of function `processInputFormat`.}

\item{subsample_num}{Number of cells in each of the training, validation, and test sets after sub-sampling.}

\item{cluster_proportion}{\itemize{
  \item \code{proportional:} (default) Same proportion of cells in each cluster, for the training and data sets, compared to the original cluster proportion.
  \item \code{equal}: Equal proportion of cells in each cluster, for the training and data sets, compared to the original cluster proportion.
}}
}
\value{
List containing
\itemize{
  \item \code{training_matrix}:
  \item \code{training_clusters}:
  \item \code{training_clusters_num}:
  \item \code{validation_matrix}:
  \item \code{validation_clusters}:
  \item \code{validation_clusters_num}:
  \item \code{test_matrix}:
  \item \code{test_clusters}:
  \item \code{test_clusters_num}:
}
}
\description{
Sub-sample and split data into training, validation and test set
}
