% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapperMethods.R
\name{sc2markerWrapper}
\alias{sc2markerWrapper}
\title{Wrapper function for sc2marker}
\usage{
sc2markerWrapper(input_matrix, clusters, num_markers = 15, ...)
}
\arguments{
\item{input_matrix}{Marker matrix (cell vs markers)}

\item{clusters}{Cell type annotation}

\item{num_markers}{Number of markers to output}
}
\value{
The most informative markers determined by sc2marker
}
\description{
Wrapper function for sc2marker
}
