% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapperMethods.R
\name{seuratWrapper}
\alias{seuratWrapper}
\title{Wrapper function for Seurat}
\usage{
seuratWrapper(input_matrix, clusters, num_markers = 15, method = "bimod", ...)
}
\arguments{
\item{input_matrix}{Marker matrix (cell vs markers)}

\item{clusters}{Cell type annotation}

\item{num_markers}{Number of markers to output}

\item{method}{Differential algorithm used in Seruat's FindAllMarkers}
}
\value{
The most informative markers determined by Seurat, based on FC
}
\description{
Wrapper function for Seurat
}
