% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapperPerformance.R
\name{xgboostPredictCell}
\alias{xgboostPredictCell}
\title{Wrapper function for xgboostPerformance}
\usage{
xgboostPredictCell(
  markers_sel,
  input_matrix_train,
  input_matrix_test,
  clusters_num_train,
  clusters_num_test,
  clusters_train,
  clusters_test,
  unique_clusters_sample,
  ...
)
}
\arguments{
\item{markers_sel}{Single cell experiment object}

\item{input_matrix_train}{Feature training matrix with cells as columns, and features as rows.}

\item{input_matrix_test}{Feature test matrix with cells as columns, and features as rows.}

\item{clusters_num_train}{Cluster annotation for training set (numerical)}

\item{clusters_num_test}{Cluster annotation for test set (numerical)}

\item{clusters_train}{Cluster annotation for training set}

\item{clusters_test}{Cluster annotation for test set}

\item{unique_clusters_sample}{Unique clusters}
}
\value{
The performance of the input markers, as determined by a xgBoost algorithm
}
\description{
Wrapper function for xgboostPerformance
}
