% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapperMethods.R
\name{xgBoostWrapper}
\alias{xgBoostWrapper}
\title{Wrapper function for xgBoost}
\usage{
xgBoostWrapper(
  input_matrix,
  clusters,
  num_markers,
  nrounds = 1500,
  nthread = 6,
  ...
)
}
\arguments{
\item{input_matrix}{Marker matrix with cells as rows, and features as columns.}

\item{clusters}{Cell type annotation}

\item{num_markers}{Number of markers to output}
}
\value{
The most informative markers determined by xgBoost
}
\description{
Wrapper function for xgBoost
}
