% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapperMethods.R
\name{citeFuseWrapper}
\alias{citeFuseWrapper}
\title{Wrapper function for CiteFuse}
\usage{
citeFuseWrapper(sce, num_markers = 15, min_freq = 3, ...)
}
\arguments{
\item{sce}{Single cell experiment object}

\item{num_markers}{Number of markers to output}
}
\value{
The most informative markers determined by CiteFuse
}
\description{
Wrapper function for CiteFuse
}
